Fix scio-repl scala specific repl settings
